@yiguolei 2022-05-06
@qzsee Have you test vectorized engine? I find you modified the non-vectorized engine.